Queue reference method (#5461)
* Queue reference method

Adds a static `get` method to the `Queue` class so that, like other lower-level SST components, one can retrieve a reference to an SQS queue from outside the app/stage. Complements the existing static `subscribe` method that allows one to subscribe to an outside SQS queue without having a reference to it. Inspired by the static `get` method in the `SnsTopic` component.

* sync

---------

Co-authored-by: Frank <frank@sst.dev>